開発: 修正: yarn start のために yarn install --cwd bin を不要にする
