+++ bin/release/scripts/uploadArtifacts.js
@@ -6,7 +6,7 @@ const { S3, S3Client } = require('@aws-sdk/client-s3');
 const { Upload } = require('@aws-sdk/lib-storage');
 const sh = require('shelljs');
 const ProgressBar = require('progress');
-const { info, error } = require('./prompt');
+const { info, error } = require('./log');
 
 /**
  *
